Initial Startup Screen
There is now a startupscreen that has a button to click called 'learn recipe' this button will then launch the game into the playerview. Additionally, the infrastructure for the recipe cards has been layed out. There will be a few options for the user to choose from. They will be able to select these recipe cards and learn a different recipe each time they play.